requirements.txt
flask>=3.0.0
authlib[flask-client]>=1.3.0
python-dotenv>=1.0.0
requests>=2.28.0
vercel_blob>=0.4.0
